+++ src/directives/VarDirective.directive.ts
@@ -3,8 +3,8 @@ import {
 } from "@angular/core"
 
 @Directive({
-  selector: '[var]',
-  exportAs: 'var'
+  selector: "[var]",
+  exportAs: "var"
 }) export class VarDirective {
   @Input() var:any;
   @Output() changed:EventEmitter<any> = new EventEmitter()